refactor : name for tsx
