+++ src/App.tsx
@@ -11,26 +11,26 @@ function App() {
   const [doneToDoList, setDoneToDoList] = useState([]);
 
   const Container = styled.div
-      `
-      text-align: center;
-      width: 360px;
-      height: 600px;
-      background-color: white;
-      display: flex;
-      flex-direction: column;
-      border-bottom: 1px solid grey;
-      border-radius: 9px;
   `
-    ;
+  text-align: center;
+  width: 360px;
+  height: 600px;
+  background-color: white;
+  display: flex;
+  flex-direction: column;
+  border-bottom: 1px solid grey;
+  border-radius: 9px;
+`
+;
 
-    const InputBox = styled.header
-    `
-    font-family: 'SuncheonB';
-    flex: 0.3;
-    border-bottom: 1px solid grey;
-    
-    `
-    ;
+const InputBox = styled.header
+`
+font-family: 'SuncheonB';
+flex: 0.3;
+border-bottom: 1px solid grey;
+
+`
+;
 
   return (
     <div>
