--- src/Finishhw.js
@@ -1,17 +1,0 @@
-import {useEffect} from "react";
-
-/*function Finishhw(){
-    if( === 0)
-  {
-    return 
-       (
-         <div>
-           과제 끝!
-         </div>
-       );
-
-  }
-  };
-  
-  export default Finishhw;
-  */
