+++ tsconfig.json
@@ -8,5 +8,5 @@
       "lib": ["ES2015", "DOM", "DOM.Iterable"]
     },
     "include": ["./src/**/*"],
-    "exclude": ["node_modules", "dist"]
+    //"exclude": ["node_modules", "dist"]
   }
